Separate visual and app layer in drafts
