+++ src/orders/components/OrderDraftListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./OrderDraftListPage";
 export * from "./OrderDraftListPage";
+export * from "./filters";
