--- src/orders/types.ts
@@ -1,6 +1,0 @@
-import { FilterOpts, MinMax } from "@saleor/types";
-
-export interface OrderDraftListFilterOpts {
-  created: FilterOpts<MinMax>;
-  customer: FilterOpts<string>;
-}
