+++ src/orders/views/OrderDraftList/OrderDraftList.tsx
@@ -46,7 +46,7 @@ import {
   saveFilterTab,
   getFilterQueryParam,
   getFilterOpts
-} from "./filter";
+} from "./filters";
 import { getSortQueryVariables } from "./sort";
 
 interface OrderDraftListProps {
