--- src/orders/views/OrderDraftList/messages.ts
@@ -1,14 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  customer: {
-    defaultMessage: "Customer",
-    description: "draft order"
-  },
-  created: {
-    defaultMessage: "Created",
-    description: "draft order"
-  }
-});
-
-export default messages;
